yield_curve_db/pyproject.toml
[project]
name = "yield-curve-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "questdb>=3.0.0",
    "requests>=2.32.4",
    "retrying>=1.4.1",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
]
